requirements.txt
torch
torchvision
PyYAML
scikit-learn
dvc
